Feat: totCnt Long -> Integer 형식 변경

+++ src/main/java/com/cineinfo/v1/dto/kofic/response/movie_list/MovieListResultRes.java
@@ -10,7 +10,7 @@ import java.util.List;
 @AllArgsConstructor
 public class MovieListResultRes {
 
-    private Long totCnt;
+    private Integer totCnt;
     private String source;
     private List<MovieListRes> movieList;
